requirements.txt
google-generativeai
gradio
PyMuPDF